fix: 修复密码登录无法点击，个人信息获取失败

+++ app/src/main/java/com/lxy/studyroom/logic/Repository.kt
@@ -1,22 +1,23 @@
 package com.lxy.studyroom.logic
 
 import android.content.Intent
-import android.widget.Toast
 import androidx.lifecycle.liveData
 import com.lxy.studyroom.StudyRoomApplication
 import com.lxy.studyroom.constant.CommonConstant
 import com.lxy.studyroom.constant.UserConstant
 import com.lxy.studyroom.enums.ResponseEnum
-import com.lxy.studyroom.extension.toast
 import com.lxy.studyroom.logic.dto.StudyRecordDTO
-import com.lxy.studyroom.logic.model.*
+import com.lxy.studyroom.logic.model.Ranking
+import com.lxy.studyroom.logic.model.RoomRecord
+import com.lxy.studyroom.logic.model.UserRank
+import com.lxy.studyroom.logic.model.UserStatistics
+import com.lxy.studyroom.logic.model.base.ResponseResult
 import com.lxy.studyroom.ui.login.LoginActivity
 import com.lxy.studyroom.util.DataStoreUtil
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
 import kotlinx.coroutines.coroutineScope
 import okhttp3.MultipartBody
-import java.lang.Exception
 
 object Repository {
 
@@ -24,19 +25,20 @@ object Repository {
      * 高阶函数，把网络请求放在 IO 线程， 统一 try catch 处理，返回LiveData
      */
     private fun <T> handleResponse(block: suspend () -> ResponseResult<T>) = run {
-        liveData(Dispatchers.IO){
+        liveData(Dispatchers.IO) {
             var result: ResponseResult<T>
             try {
                 result = block()
-                if (result.code == ResponseEnum.NEED_LOGIN.code){
+                if (result.code == ResponseEnum.NEED_LOGIN.code) {
 
                     saveUserLoginStatus(false)
-                    val intent = Intent(StudyRoomApplication.context, LoginActivity::class.java).apply {
-                        flags = Intent.FLAG_ACTIVITY_NEW_TASK
-                    }
+                    val intent =
+                        Intent(StudyRoomApplication.context, LoginActivity::class.java).apply {
+                            flags = Intent.FLAG_ACTIVITY_NEW_TASK
+                        }
                     StudyRoomApplication.context.startActivity(intent)
                 }
-            }catch (e: Exception){
+            } catch (e: Exception) {
                 e.printStackTrace()
                 result = ResponseResult()
                 result.code = ResponseEnum.DATA_ERROR.code
@@ -60,7 +62,7 @@ object Repository {
     /**
      * 图书馆详情
      */
-    fun getClassifyDetail(classifyId: Int) = handleResponse {
+    fun getClassifyDetail(classifyId: Long) = handleResponse {
         StudyRoomNetwork.getClassifyDetail(classifyId)
     }
 
@@ -74,14 +76,14 @@ object Repository {
     /**
      * 结束自习
      */
-    fun stopStudy(recordId: Int) = handleResponse {
+    fun stopStudy(recordId: Long) = handleResponse {
         StudyRoomNetwork.stopStudy(recordId)
     }
 
     /**
      * 学习记录详情
      */
-    fun getLearningRecordDetail(recordId: Int) = handleResponse {
+    fun getLearningRecordDetail(recordId: Long) = handleResponse {
         StudyRoomNetwork.getLearningRecordDetail(recordId)
     }
 
@@ -92,28 +94,28 @@ object Repository {
     /**
      * 提交学习时长
      */
-    fun submitStudyDuration(recordId: Int,duration: Int) = handleResponse {
+    fun submitStudyDuration(recordId: Long, duration: Int) = handleResponse {
         StudyRoomNetwork.submitStudyDuration(recordId, duration)
     }
 
     /**
      * 获取笔记详情
      */
-    fun getStudyNoteDetail(recordId: Int) = handleResponse {
+    fun getStudyNoteDetail(recordId: Long) = handleResponse {
         StudyRoomNetwork.getStudyNoteDetail(recordId)
     }
 
     /**
      * 获取笔记列表
      */
-    fun getStudyNotes(page: Int,limit: Int) = handleResponse {
+    fun getStudyNotes(page: Int, limit: Int) = handleResponse {
         StudyRoomNetwork.getStudyNotes(page, limit)
     }
 
     /**
      * 编辑笔记
      */
-    fun saveStudyNote(recordId: Int,content: String,pic: String) = handleResponse {
+    fun saveStudyNote(recordId: Long, content: String, pic: String) = handleResponse {
         StudyRoomNetwork.saveStudyNote(recordId, content, pic)
     }
 
@@ -121,22 +123,22 @@ object Repository {
         StudyRoomNetwork.getSelfRankings()
     }
 
-    fun removeStudyNote(recordId: Int) = handleResponse {
+    fun removeStudyNote(recordId: Long) = handleResponse {
         StudyRoomNetwork.removeStudyNote(recordId)
     }
 
-    fun getStudyRecord(page: Int,limit: Int) = handleResponse {
+    fun getStudyRecord(page: Int, limit: Int) = handleResponse {
         StudyRoomNetwork.getStudyRecord(page, limit)
     }
 
-    fun upload(file :List<MultipartBody.Part>) = handleResponse {
+    fun upload(file: List<MultipartBody.Part>) = handleResponse {
         StudyRoomNetwork.upload(file)
     }
 
     /**
      * 利用协程，并发执行俩请求提高效率
      */
-    fun getRoomDetailAndRecords(roomId: Int) = handleResponse {
+    fun getRoomDetailAndRecords(roomId: Long) = handleResponse {
         coroutineScope {
             val result: ResponseResult<RoomRecord>
 
@@ -150,11 +152,11 @@ object Repository {
             val roomResponse = room.await()
             val recordResponse = record.await()
 
-            if (roomResponse.isSuccess() && recordResponse.isSuccess()){
+            if (roomResponse.isSuccess() && recordResponse.isSuccess()) {
                 result = ResponseResult()
                 result.code = ResponseEnum.SUCCESS.code
-                result.data = RoomRecord(roomResponse.data!!,recordResponse.data)
-            }else{
+                result.data = RoomRecord(roomResponse.data!!, recordResponse.data)
+            } else {
                 throw RuntimeException("数据不对头")
             }
 
@@ -162,8 +164,8 @@ object Repository {
         }
     }
 
-    fun getUserRecordsAndStatistics(page: Int,limit: Int)= handleResponse {
-        coroutineScope{
+    fun getUserRecordsAndStatistics(page: Int, limit: Int) = handleResponse {
+        coroutineScope {
             val result: ResponseResult<UserStatistics>
 
             val user = async {
@@ -175,18 +177,17 @@ object Repository {
 
             val userResp = user.await()
             val recordsResp = records.await()
-            if (userResp.isSuccess() && recordsResp.isSuccess()){
+            if (userResp.isSuccess() && recordsResp.isSuccess()) {
                 result = ResponseResult()
                 result.code = ResponseEnum.SUCCESS.code
-                result.data = UserStatistics(userResp.data!!,recordsResp.data)
-            }else{
+                result.data = UserStatistics(userResp.data!!, recordsResp.data)
+            } else {
                 throw RuntimeException("数据不对头")
             }
 
             result
         }
     }
-
 
 
     /**
@@ -206,13 +207,13 @@ object Repository {
             val rankResponse = rank.await()
             val ruleResponse = rule.await()
 
-            if (rankResponse.isSuccess() && ruleResponse.isSuccess()){
+            if (rankResponse.isSuccess() && ruleResponse.isSuccess()) {
                 result = ResponseResult()
                 val rankList = rankResponse.data ?: ArrayList<UserRank>()
                 val ruleData = ruleResponse.data ?: ""
                 result.code = ResponseEnum.SUCCESS.code
-                result.data = Ranking(rankList,ruleData)
-            }else{
+                result.data = Ranking(rankList, ruleData)
+            } else {
                 throw RuntimeException("数据不对头")
             }
 
@@ -223,15 +224,15 @@ object Repository {
     /**
      * 反馈
      */
-    fun submitFeedBack(content: String,pic: String) = handleResponse {
+    fun submitFeedBack(content: String, pic: String) = handleResponse {
         StudyRoomNetwork.submitFeedBack(content, pic)
     }
 
-    fun getFeedBackList(page: Int,limit: Int,mine: Int) = handleResponse {
+    fun getFeedBackList(page: Int, limit: Int, mine: Int) = handleResponse {
         StudyRoomNetwork.getFeedBackList(page, limit, mine)
     }
 
-    fun getFeedBackDetail(id: Int) = handleResponse {
+    fun getFeedBackDetail(id: Long) = handleResponse {
         StudyRoomNetwork.getFeedBackDetail(id)
     }
 
@@ -240,15 +241,21 @@ object Repository {
         StudyRoomNetwork.getUserInfo()
     }
 
-    fun getUserInfoById(userId: Int) = handleResponse {
+    fun getUserInfoById(userId: Long) = handleResponse {
         StudyRoomNetwork.getUserInfoById(userId)
     }
 
-    fun updatePassword(phone: String,verificationCode: String,password: String) = handleResponse {
+    fun updatePassword(phone: String, verificationCode: String, password: String) = handleResponse {
         StudyRoomNetwork.updatePassword(phone, verificationCode, password)
     }
 
-    fun updateUserInfo(name: String,gender: String,address: String,profilePath: String,coverPath: String) = handleResponse {
+    fun updateUserInfo(
+        name: String,
+        gender: String,
+        address: String,
+        profilePath: String,
+        coverPath: String
+    ) = handleResponse {
         StudyRoomNetwork.updateUserInfo(name, gender, address, profilePath, coverPath)
     }
 
@@ -256,11 +263,11 @@ object Repository {
         StudyRoomNetwork.getVerificationCode(phone)
     }
 
-    fun loginByVerificationCode(phone: String,verificationCode: String) = handleResponse {
+    fun loginByVerificationCode(phone: String, verificationCode: String) = handleResponse {
         StudyRoomNetwork.loginByVerificationCode(phone, verificationCode)
     }
 
-    fun login(phone: String,password: String) = handleResponse {
+    fun login(phone: String, password: String) = handleResponse {
         StudyRoomNetwork.login(phone, password)
     }
 
@@ -273,39 +280,39 @@ object Repository {
      */
     fun getUserLoginStatus(): Boolean = DataStoreUtil.getBoolean(UserConstant.LOGIN_STATUS, false)
 
-    fun saveUserLoginStatus(flag: Boolean) = DataStoreUtil.putBoolean(UserConstant.LOGIN_STATUS,flag)
+    fun saveUserLoginStatus(flag: Boolean) =
+        DataStoreUtil.putBoolean(UserConstant.LOGIN_STATUS, flag)
 
-    fun saveToken(token: String) = DataStoreUtil.putString(CommonConstant.BASE_TOKEN_NAME,token)
+    fun saveToken(token: String) = DataStoreUtil.putString(CommonConstant.BASE_TOKEN_NAME, token)
 
     /**
      * 用户协议与隐私政策是否被选中
      */
     fun getAgreePolicy(): Boolean = DataStoreUtil.getBoolean(UserConstant.AGREE_POLICY, false)
 
-    fun saveAgreePolicy(flag: Boolean) = DataStoreUtil.putBoolean(UserConstant.AGREE_POLICY,flag)
+    fun saveAgreePolicy(flag: Boolean) = DataStoreUtil.putBoolean(UserConstant.AGREE_POLICY, flag)
 
     /**
      * 用户手机号
      */
-    fun getUserPhone(): String = DataStoreUtil.getString(UserConstant.PHONE,"")
+    fun getUserPhone(): String = DataStoreUtil.getString(UserConstant.PHONE, "")
 
-    fun saveUserPhone(phone: String) = DataStoreUtil.putString(UserConstant.PHONE,phone)
+    fun saveUserPhone(phone: String) = DataStoreUtil.putString(UserConstant.PHONE, phone)
 
     /**
      * 用户id
      */
-    fun getUserId(): Int = DataStoreUtil.getInt(UserConstant.ID,-1)
+    fun getUserId(): Long = DataStoreUtil.getLong(UserConstant.ID, -1L)
 
-    fun saveUserId(userId: Int) = DataStoreUtil.putInt(UserConstant.ID,userId)
+    fun saveUserId(userId: Long) = DataStoreUtil.putLong(UserConstant.ID, userId)
 
     /**
      * 用户密码
      */
-    fun getUserPassword(): String = DataStoreUtil.getString(UserConstant.PASSWORD,"")
+    fun getUserPassword(): String = DataStoreUtil.getString(UserConstant.PASSWORD, "")
 
-    fun saveUserPassword(password: String) = DataStoreUtil.putString(UserConstant.PASSWORD,password)
-
-
+    fun saveUserPassword(password: String) =
+        DataStoreUtil.putString(UserConstant.PASSWORD, password)
 
 
 }
