+++ app/src/main/java/com/lxy/studyroom/logic/dto/StudyDurationDTO.kt
@@ -6,4 +6,4 @@ import java.io.Serializable
  * 学习时长数据传输对象
  */
 
-data class StudyDurationDTO(val duration: Int, val recordId: Int):Serializable
+data class StudyDurationDTO(val duration: Int, val recordId: Long):Serializable
