+++ app/src/main/java/com/lxy/studyroom/logic/model/LibraryRoomDetail.kt
@@ -3,8 +3,8 @@ package com.lxy.studyroom.logic.model
 import java.io.Serializable
 
 data class LibraryRoomDetail(
-    val id: Int,
-    val parentId: Int,
+    val id: Long,
+    val parentId: Long,
     val name: String,
     val personCount: Int,
     val parentName: String,
