+++ app/src/main/java/com/lxy/studyroom/logic/model/RoomDetail.kt
@@ -3,9 +3,9 @@ package com.lxy.studyroom.logic.model
 import java.io.Serializable
 
 data class RoomDetail(
-    val catalogId: Int,
-    val classifyId: Int,
-    val parentId: Int,
+    val catalogId: Long,
+    val classifyId: Long,
+    val parentId: Long,
     val level: Int,
     val catalogName: String,
     val personCount: Int,
