+++ app/src/main/java/com/lxy/studyroom/logic/network/service/StudyRecordService.kt
@@ -1,6 +1,6 @@
 package com.lxy.studyroom.logic.network.service
 
-import com.lxy.studyroom.logic.model.ResponseResult
+import com.lxy.studyroom.logic.model.base.ResponseResult
 import com.lxy.studyroom.logic.model.StudyRecord
 import retrofit2.Call
 import retrofit2.http.POST
@@ -18,19 +18,19 @@ interface StudyRecordService {
      * 获取自习笔记详情
      */
     @POST("studyRecord/getStudyNoteDetail")
-    fun getStudyNoteDetail(@Query("recordId") recordId: Int) : Call<ResponseResult<StudyRecord?>>
+    fun getStudyNoteDetail(@Query("recordId") recordId: Long) : Call<ResponseResult<StudyRecord?>>
 
     /**
      * 编辑自习笔记
      */
     @POST("studyRecord/saveStudyNote")
-    fun saveStudyNote(@Query("recordId") recordId: Int,@Query("content") content: String,@Query("pic") pic: String) : Call<ResponseResult<String?>>
+    fun saveStudyNote(@Query("recordId") recordId: Long,@Query("content") content: String,@Query("pic") pic: String) : Call<ResponseResult<String?>>
 
     /**
      * 删除自习笔记
      */
     @POST("studyRecord/removeStudyNote")
-    fun removeStudyNote(@Query("recordId") recordId: Int) : Call<ResponseResult<String?>>
+    fun removeStudyNote(@Query("recordId") recordId: Long) : Call<ResponseResult<String?>>
 
     /**
      * 获取自习记录
